requirements.txt
# Backend Dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
requests==2.31.0
Pillow==10.1.0
django-filter==23.4
djangorestframework-simplejwt==5.3.0

# API Integrations
twilio==8.10.0
vonage==3.12.0

# AI/ML Dependencies
tensorflow==2.15.0
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
huggingface-hub==0.19.4
opencv-python==4.8.1.78

# Development
django-debug-toolbar==4.2.0
pytest-django==4.7.0
black==23.11.0
flake8==6.1.0

# Deployment
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
